properly use mise for claude code

+++ migrate.sh
@@ -3,17 +3,21 @@ set -eo pipefail -ux
 
 # claude
 
-for DIR in \
-  ~/.claude \
-  ~/.cache/claude \
-  ~/.cache/claude-cli-nodejs \
-  ~/.local/bin/claude \
-  ~/.local/share/claude \
-  ~/.local/state/claude; do
-  rm -rf "$DIR"
-done
+if [[ $HOST == 'worker' ]]; then
+  for DIR in \
+    ~/.claude \
+    ~/.cache/claude \
+    ~/.cache/claude-cli-nodejs \
+    ~/.local/bin/claude \
+    ~/.local/share/claude \
+    ~/.local/state/claude; do
+    rm -rf "$DIR"
+  done
 
-pushd ~/code/dot && git pull && popd && mise install
+  pushd ~/code/dot && git pull && popd
+  "${BASH_SOURCE%/*}"/claude.sh
+
+fi
 
 # fonts
